requirements.txt
requests>=2.28.0
python-dateutil>=2.8.2
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
